requirements.txt
 asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
imutils==0.5.4
intel-openmp==2021.4.0
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
open3d==0.18.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pooch==1.8.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
pytz==2024.1
pyvista==0.43.8
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.0
scooby==0.10.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
thop==0.1.1.post2209072238
tifffile==2024.5.10
tk==0.1.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.15
urllib3==2.2.1
vtk==9.3.0
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
zipp==3.18.2
zope.interface==6.3